style: rename test files
