+++ package.json
@@ -7,6 +7,7 @@
   },
   "devDependencies": {
     "@types/node": "^20",
+    "@types/semver": "^7.5.8",
     "@vercel/ncc": "^0.38.0",
     "typescript": "^5",
     "vitest": "^3.0.0"
